refactor: apply SOLID principles on product card footer
