+++ app/products/[...slug]/page.tsx
@@ -1,5 +1,5 @@
 import { BreadCrumb } from '@/components/bread-crumb';
-import { ProductCardActions } from '@/components/product-card-actions';
+import { ProductCardFooter } from '@/components/product-card-footer';
 import { Category } from '@/containers/routes/single-product/category';
 import { Images } from '@/containers/routes/single-product/images';
 import { MiniAttributes } from '@/containers/routes/single-product/mini-attributes';
@@ -40,7 +40,7 @@ export default async function Page(props: IPageProps) {
                 priceWithoutDiscount={data.priceWithoutDiscount}
               />
               <Quantity quantity={data.quantity} />
-              <ProductCardActions type="single-product" data={data} />
+              <ProductCardFooter type="single-product" data={data} />
             </div>
             <div className="mt-5 flex w-full flex-col gap-5 sm:flex-row-reverse md:mt-0 xl:gap-5">
               <MiniAttributes attributes={data.attributes} />
