+++ containers/routes/home/hero-offer-slider.tsx
@@ -8,8 +8,9 @@ import { useEffect, useRef, useState } from 'react';
 import { HiChevronLeft, HiChevronRight } from 'react-icons/hi2';
 import { Autoplay } from 'swiper/modules';
 import { Swiper, SwiperSlide } from 'swiper/react';
-import { ProductCardActions } from '@/components/product-card-actions';
+import { ProductCardFooter } from '@/components/product-card-footer';
 import { productsData } from '@/resources/products';
+import { TProduct } from '@/types/product';
 import { cn } from '@/utils/cn';
 import { shuffleArray } from '@/utils/shuffle-array';
 
@@ -59,27 +60,12 @@ export function HeroOfferSlider() {
                       {item.title.fa}
                     </p>
                   </Link>
-                  {/* price */}
-                  <div>
-                    <del className="absolute bottom-8 left-20 text-smp text-gray-50">
-                      {item.priceWithDiscount.toLocaleString('fa-IR')}
-                    </del>
-                    <div>
-                      <p className="absolute bottom-1 left-8 text-2xl font-bold text-white">
-                        {item.priceWithoutDiscount.toLocaleString('fa-IR')}
-                      </p>
-                      <p className="absolute bottom-5 left-2 -rotate-90 text-xs font-bold text-white">
-                        تومان
-                      </p>
-                      <div className="absolute bottom-9 left-8 flex h-[22px] gap-1 rounded-full rounded-bl-none bg-white px-2">
-                        <p className="pt-0.5 font-bold text-red">
-                          {item.discount}
-                        </p>
-                        <p className="pt-1 text-sm font-bold text-red">%</p>
-                      </div>
-                    </div>
-                  </div>
-                  <ProductCardActions data={item} type="offer-slider" />
+                  {/* price/actions */}
+                  <ProductCardFooter
+                    data={item}
+                    type="offer-slider"
+                    priceComponent={<OfferSliderCardPrice item={item} />}
+                  />
                 </div>
               </SwiperSlide>
             );
@@ -88,6 +74,28 @@ export function HeroOfferSlider() {
       </div>
       <NavigationMobile swiperRef={swiperRef} />
     </section>
+  );
+}
+
+function OfferSliderCardPrice({ item }: { item: TProduct }) {
+  return (
+    <div>
+      <del className="absolute bottom-8 left-20 text-smp text-gray-50">
+        {item.priceWithDiscount.toLocaleString('fa-IR')}
+      </del>
+      <div>
+        <p className="absolute bottom-1 left-8 text-2xl font-bold text-white">
+          {item.priceWithoutDiscount.toLocaleString('fa-IR')}
+        </p>
+        <p className="absolute bottom-5 left-2 -rotate-90 text-xs font-bold text-white">
+          تومان
+        </p>
+        <div className="absolute bottom-9 left-8 flex h-[22px] gap-1 rounded-full rounded-bl-none bg-white px-2">
+          <p className="pt-0.5 font-bold text-red">{item.discount}</p>
+          <p className="pt-1 text-sm font-bold text-red">%</p>
+        </div>
+      </div>
+    </div>
   );
 }
